(Hopefully) fixed repeating number issue and updated field value
The repeating number issue, which occurred in item types Children,
Toddlers, Infants, and Pets seemed to have occurred because of a
variable type error (string instead of int). All values are cast as int
now, which appears to have solved the issue.

Also, changed the field label in Add Checklist from Item Type to
Checklist Type.